+++ change_log.md
@@ -1,2 +1,3 @@
-- Journal main screen (still header bug)
-- fab structured better
+- not showing tabnavigatorin addMood screen
+- info boxes
+- adding space to fab so that all elements can be seen when scrolling down